tests/wpt/tests/html/canvas/element/canvas-host/2d.canvas.host.size.attributes.idl.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.canvas.host.size.attributes.idl</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.host.size.attributes.idl</h1>
<p class="desc">Getting/setting width/height IDL attributes</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Getting/setting width/height IDL attributes");
_addTest(function(canvas, ctx) {

  canvas.width = '100';
  canvas.height = '100';
  _assertSame(canvas.width, 100, "canvas.width", "100");
  _assertSame(canvas.height, 100, "canvas.height", "100");
  canvas.width = 301.999;
  canvas.height = 301.001;
  _assertSame(canvas.width, 301, "canvas.width", "301");
  _assertSame(canvas.height, 301, "canvas.height", "301");
  canvas.width = "+1.5e2";
  canvas.height = "0x96";
  _assertSame(canvas.width, 150, "canvas.width", "150");
  _assertSame(canvas.height, 150, "canvas.height", "150");

});
</script>
